requirements.txt
flask
torch
torchvision
matplotlib
scikit-image
pillow
blend-modes
numpy
ibm-cos-sdk
opencv-python
